+++ jwskate/jwa/signature/__init__.py
@@ -1,11 +1,14 @@
 """This module exposes all the Signature algorithms available from `jwskate`."""
 
-from .ec import ES256, ES256K, ES384, ES512
+from .ec import ES256, ES256K, ES384, ES512, BaseECSignatureAlg
 from .eddsa import EdDsa
-from .hmac import HS256, HS384, HS512
-from .rsa import PS256, PS384, PS512, RS256, RS384, RS512
+from .hmac import HS256, HS384, HS512, BaseHMACSigAlg
+from .rsa import PS256, PS384, PS512, RS256, RS384, RS512, BaseRSASigAlg
 
 __all__ = [
+    "BaseECSignatureAlg",
+    "BaseHMACSigAlg",
+    "BaseRSASigAlg",
     "ES256",
     "ES256K",
     "ES384",
