+++ jwskate/jwk/ec.py
@@ -7,7 +7,7 @@ from typing import Any, List, Mapping, Optional, Union
 
 from backports.cached_property import cached_property
 from binapy import BinaPy
-from cryptography.hazmat.primitives import asymmetric, serialization
+from cryptography.hazmat.primitives import asymmetric
 from cryptography.hazmat.primitives.asymmetric import ec
 
 from jwskate.jwa import (
