+++ jwskate/jws/signature.py
@@ -7,7 +7,7 @@ from typing import Any, Dict, Iterable, Mapping, Optional, Type, TypeVar, Union
 from backports.cached_property import cached_property
 from binapy import BinaPy
 
-from jwskate.jwk.base import Jwk, to_jwk
+from jwskate.jwk import Jwk, to_jwk
 from jwskate.token import BaseJsonDict
 
 S = TypeVar("S", bound="JwsSignature")
